+++ app/src/main/java/org/thoughtcrime/securesms/ApplicationContext.java
@@ -86,7 +86,7 @@ import org.thoughtcrime.securesms.sskenvironment.ProfileManager;
 import org.thoughtcrime.securesms.sskenvironment.ReadReceiptManager;
 import org.thoughtcrime.securesms.sskenvironment.TypingStatusRepository;
 import org.thoughtcrime.securesms.util.Broadcaster;
-import org.thoughtcrime.securesms.util.VersionUtil;
+import org.thoughtcrime.securesms.util.VersionDataFetcher;
 import org.thoughtcrime.securesms.util.dynamiclanguage.LocaleParseHelper;
 import org.thoughtcrime.securesms.webrtc.CallMessageProcessor;
 import org.webrtc.PeerConnectionFactory;
@@ -111,7 +111,6 @@ import javax.inject.Inject;
 import dagger.hilt.EntryPoints;
 import dagger.hilt.android.HiltAndroidApp;
 import kotlin.Unit;
-import kotlinx.coroutines.Job;
 import network.loki.messenger.BuildConfig;
 import network.loki.messenger.libsession_util.ConfigBase;
 import network.loki.messenger.libsession_util.UserProfile;
@@ -143,7 +142,6 @@ public class ApplicationContext extends Application implements DefaultLifecycleO
     private HandlerThread conversationListHandlerThread;
     private Handler conversationListHandler;
     private PersistentLogger persistentLogger;
-    private VersionUtil versionUtil;
 
     @Inject LokiAPIDatabase lokiAPIDatabase;
     @Inject public Storage storage;
@@ -153,6 +151,7 @@ public class ApplicationContext extends Application implements DefaultLifecycleO
     @Inject PushRegistry pushRegistry;
     @Inject ConfigFactory configFactory;
     @Inject LastSentTimestampCache lastSentTimestampCache;
+    @Inject VersionDataFetcher versionDataFetcher;
     CallMessageProcessor callMessageProcessor;
     MessagingModuleConfiguration messagingModuleConfiguration;
 
@@ -250,7 +249,6 @@ public class ApplicationContext extends Application implements DefaultLifecycleO
         resubmitProfilePictureIfNeeded();
         loadEmojiSearchIndexIfNeeded();
         EmojiSource.refresh();
-        versionUtil = new VersionUtil(textSecurePreferences);
 
         NetworkConstraint networkConstraint = new NetworkConstraint.Factory(this).create();
         HTTP.INSTANCE.setConnectedToNetwork(networkConstraint::isMet);
@@ -279,7 +277,7 @@ public class ApplicationContext extends Application implements DefaultLifecycleO
         });
 
         // fetch last version data
-        versionUtil.startTimedVersionCheck();
+        versionDataFetcher.startTimedVersionCheck();
     }
 
     @Override
@@ -292,14 +290,14 @@ public class ApplicationContext extends Application implements DefaultLifecycleO
             poller.stopIfNeeded();
         }
         ClosedGroupPollerV2.getShared().stopAll();
-        versionUtil.stopTimedVersionCheck();
+        versionDataFetcher.stopTimedVersionCheck();
     }
 
     @Override
     public void onTerminate() {
         stopKovenant(); // Loki
         OpenGroupManager.INSTANCE.stopPolling();
-        versionUtil.clear();
+        versionDataFetcher.clear();
         super.onTerminate();
     }
